requirements.txt
aiosqlite==0.20.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.30.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
coverage==7.13.4
fastapi==0.115.0
greenlet==3.1.1
grpcio==1.68.0
grpcio-tools==1.68.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
idna==3.11
iniconfig==2.3.0
kagglehub==0.3.4
mako==1.3.10
markupsafe==3.0.3
mypy==1.13.0
mypy-extensions==1.1.0
packaging==26.0
pika==1.3.2
pip==26.0.1
pluggy==1.6.0
protobuf==5.28.3
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-settings==2.6.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==7.0.0
python-dotenv==1.0.1
pyyaml==6.0.3
redis==5.1.0
requests==2.32.5
ruff==0.6.9
setuptools==82.0.0
sniffio==1.3.1
sqlalchemy==2.0.36
starlette==0.38.6
structlog==24.4.0
tqdm==4.67.3
typing-extensions==4.15.0
urllib3==2.6.3
uvicorn==0.30.6
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
